Move dev polling to InRepoWorker
